sc2(tarfs): skip renaming layer file if it already exists
When using the tarfs package to convert an OCI image where two layers
have the same digest, there is a race condition where two snapshots,
corresponding to two different layers, want to target the same file in
the cache. It is enough to check if the file has already been created by
a previous snapshot before renaming.

Note that we still need to do some work for the repeated snapshot, as we
need to create its own bootstrap layer.